requirements.txt
asgiref==3.2.10
certifi==2019.11.28
chardet==3.0.4
coverage==5.3
Django==3.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
idna==2.8
psycopg2==2.8.6
pytz==2020.1
requests==2.22.0
sqlparse==0.3.1
urllib3==1.25.3
